+++ apps/elf-eval/tests/real_world_job_benchmark/consolidation_knowledge_tests_fixtures/consolidation.rs
@@ -1,0 +1,70 @@
+use color_eyre::Result;
+use serde_json::Value;
+
+use crate::support;
+
+#[test]
+fn consolidation_fixtures_report_reviewable_proposal_metrics() -> Result<()> {
+	let report = support::run_json_report_from(support::consolidation_fixture_dir())?;
+
+	assert_eq!(report.pointer("/summary/job_count").and_then(Value::as_u64), Some(4));
+	assert_eq!(report.pointer("/summary/pass").and_then(Value::as_u64), Some(4));
+	assert_eq!(
+		report.pointer("/summary/consolidation/proposal_count").and_then(Value::as_u64),
+		Some(4)
+	);
+	assert_eq!(
+		report.pointer("/summary/consolidation/source_mutation_count").and_then(Value::as_u64),
+		Some(0)
+	);
+	assert_eq!(
+		report
+			.pointer("/summary/consolidation/proposal_unsupported_claim_count")
+			.and_then(Value::as_u64),
+		Some(1)
+	);
+	assert_eq!(
+		report.pointer("/summary/consolidation/executable_gap_count").and_then(Value::as_u64),
+		Some(0)
+	);
+	assert_eq!(
+		report.pointer("/summary/consolidation/lineage_completeness").and_then(Value::as_f64),
+		Some(1.0)
+	);
+	assert_eq!(
+		report.pointer("/summary/consolidation/review_action_correctness").and_then(Value::as_f64),
+		Some(1.0)
+	);
+
+	let jobs = support::array_at(&report, "/jobs")?;
+	let project_summary =
+		support::find_by_field(jobs, "/job_id", "consolidation-project-summary-apply-001")?;
+	let contradiction =
+		support::find_by_field(jobs, "/job_id", "consolidation-contradiction-report-discard-001")?;
+
+	assert_eq!(
+		project_summary
+			.pointer("/consolidation/proposals/0/actual_review_action")
+			.and_then(Value::as_str),
+		Some("apply")
+	);
+	assert_eq!(
+		contradiction
+			.pointer("/consolidation/proposals/0/actual_review_action")
+			.and_then(Value::as_str),
+		Some("discard")
+	);
+	assert_eq!(
+		contradiction
+			.pointer("/consolidation/proposals/0/unsupported_claim_count")
+			.and_then(Value::as_u64),
+		Some(1)
+	);
+
+	let suites = support::array_at(&report, "/suites")?;
+	let consolidation_suite = support::find_by_field(suites, "/suite_id", "consolidation")?;
+
+	assert_eq!(consolidation_suite.pointer("/status").and_then(Value::as_str), Some("pass"));
+
+	Ok(())
+}
