+++ apps/elf-eval/tests/real_world_job_benchmark/consolidation_knowledge_tests_fixtures/project_decisions.rs
@@ -1,0 +1,92 @@
+use std::fs;
+
+use color_eyre::Result;
+use serde_json::Value;
+
+use crate::support;
+
+#[test]
+fn project_decisions_fixtures_report_decision_policy_cases() -> Result<()> {
+	let report = support::run_json_report_from(support::project_decisions_fixture_dir())?;
+
+	assert_eq!(report.pointer("/summary/job_count").and_then(Value::as_u64), Some(5));
+	assert_eq!(report.pointer("/summary/pass").and_then(Value::as_u64), Some(5));
+	assert_eq!(report.pointer("/summary/wrong_result").and_then(Value::as_u64), Some(0));
+	assert_eq!(report.pointer("/summary/not_encoded").and_then(Value::as_u64), Some(0));
+	assert_eq!(
+		report.pointer("/summary/conflict_detection_count").and_then(Value::as_u64),
+		Some(2)
+	);
+	assert_eq!(
+		report.pointer("/summary/update_rationale_available_count").and_then(Value::as_u64),
+		Some(5)
+	);
+	assert_eq!(
+		report.pointer("/summary/expected_evidence_recall").and_then(Value::as_f64),
+		Some(1.0)
+	);
+
+	let suites = support::array_at(&report, "/suites")?;
+	let project_decisions = support::find_by_field(suites, "/suite_id", "project_decisions")?;
+
+	assert_eq!(project_decisions.pointer("/status").and_then(Value::as_str), Some("pass"));
+	assert_eq!(project_decisions.pointer("/encoded_job_count").and_then(Value::as_u64), Some(5));
+	assert_eq!(
+		project_decisions.pointer("/update_rationale_available_count").and_then(Value::as_u64),
+		Some(5)
+	);
+
+	let jobs = support::array_at(&report, "/jobs")?;
+	let accepted =
+		support::find_by_field(jobs, "/job_id", "project-decision-accepted-typed-failures-001")?;
+	let reversal =
+		support::find_by_field(jobs, "/job_id", "project-decision-reversal-live-baseline-001")?;
+	let validation =
+		support::find_by_field(jobs, "/job_id", "project-decision-current-validation-gate-001")?;
+	let tradeoff =
+		support::find_by_field(jobs, "/job_id", "project-decision-tradeoff-fixture-backed-001")?;
+	let caveat =
+		support::find_by_field(jobs, "/job_id", "project-decision-private-manifest-caveat-001")?;
+
+	assert_eq!(accepted.pointer("/answer_type").and_then(Value::as_str), Some("decision_record"));
+	assert_eq!(
+		accepted.pointer("/expected_evidence").and_then(Value::as_array).map(Vec::len),
+		Some(2)
+	);
+	assert_eq!(
+		reversal.pointer("/evolution/historical_evidence/0").and_then(Value::as_str),
+		Some("live-baseline-suite-win-old")
+	);
+	assert_eq!(
+		validation.pointer("/evolution/current_evidence/0").and_then(Value::as_str),
+		Some("validation-gate-current-decodex")
+	);
+	assert_eq!(tradeoff.pointer("/requires_caveat").and_then(Value::as_bool), Some(true));
+	assert_eq!(caveat.pointer("/can_answer_unknown").and_then(Value::as_bool), Some(true));
+
+	for job in jobs {
+		let expected_evidence = support::array_at(job, "/expected_evidence")?;
+
+		assert!(
+			!expected_evidence.is_empty(),
+			"project decision job {} must declare required evidence",
+			job.pointer("/job_id").and_then(Value::as_str).unwrap_or("<unknown>")
+		);
+	}
+	for entry in fs::read_dir(support::project_decisions_fixture_dir())? {
+		let path = entry?.path();
+
+		if path.extension().and_then(|ext| ext.to_str()) != Some("json") {
+			continue;
+		}
+
+		let fixture = serde_json::from_str::<Value>(&fs::read_to_string(path)?)?;
+		let required_evidence = support::array_at(&fixture, "/required_evidence")?;
+		let negative_traps = support::array_at(&fixture, "/negative_traps")?;
+
+		assert!(!required_evidence.is_empty());
+		assert!(!negative_traps.is_empty());
+	}
+
+	Ok(())
+}
