+++ apps/elf-eval/tests/real_world_job_benchmark/support/report_paths_project_files.rs
@@ -1,0 +1,9 @@
+use std::path::PathBuf;
+
+use color_eyre::Result;
+
+use crate::support;
+
+pub(crate) fn readme_path() -> Result<PathBuf> {
+	Ok(support::workspace_root()?.join("README.md"))
+}
